requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
python-dotenv
streamlit
pypdf
pysqlite3-binary
gdown
openpyxl